refactor: extract loadSvgWithFill to shared svg.utils and update icon shape and fab button components
